Seperated concerns in OperationExecution, added unit tests
Refactored OperationExecuter to seperate the output code so that's it's just creating the execution logs. Returns null if a null reference is passed in.

Program class now handles printing the execution logs in a helper method.

Added FluentAssertions to the test project for easy list comparison.

Modified the existing OperationExecuter unit test to check for the null reference that's returned.

Added OperationExecuter tests for an empty list arguement and multiple valid operations.